+++ Source/BansheeEngine/Source/BsSprite.cpp
@@ -329,9 +329,9 @@ namespace bs
 		Vector<Plane> clipPlanes =
 		{
 			Plane(Vector3(1.0f, 0.0f, 0.0f), (float)clipRect.x),
-			Plane(Vector3(-1.0f, 0.0f, 0.0f), (float)-(clipRect.x + clipRect.width)),
+			Plane(Vector3(-1.0f, 0.0f, 0.0f), (float)-(clipRect.x + (INT32)clipRect.width)),
 			Plane(Vector3(0.0f, 1.0f, 0.0f), (float)clipRect.y),
-			Plane(Vector3(0.0f, -1.0f, 0.0f), (float)-(clipRect.y + clipRect.height))
+			Plane(Vector3(0.0f, -1.0f, 0.0f), (float)-(clipRect.y + (INT32)clipRect.height))
 		};
 
 		MeshUtility::clip2D(vertices, uv, numTris, vertStride, clipPlanes, writeCallback);
